Add CMakeLists.txt for Gunz and the projects on which it depends
Removed some unused and outdated .cpp files that were getting picked up by the file glob but weren't in the solutions, and moved some .cpp files that were accidentally in /Include to /Source.

It's missing tests stuff, missing CMakeLists for the rest of the solutions, and it doesn't output the Gunz binaries in a single folder (like runtime), but in a folder for each solution configuration in the output path.