Fix a bug where hitting accelerator keys could activate buttons while selecting action keys in options
For instance, trying to bind a key to 'C' but accidentally activating the cancel button and closing the options screen

MWidget::EventAccelerator was returning only if the focused widget's most derived type was MEdit, i.e. if it actually was exactly an MEdit, which failed for ZActionKey because it's a subclass of MEdit. This commit changes it to return when the focused widget is any kind of MEdit, determined via dynamic_cast.