Fix bug where, in bordered windowed mode, NewChat would get the cursor position wrong
The code was using the Windows API GetCursorPos, which gets the global/screen position, i.e. the position over the ENTIRE screen, rather than the position within the client area of the Gunz window (the basis of which can be translated by moving the window, and is already translated by the addition of the border space to the window frame).

The new code uses the MEvent::LatestPos value instead, which is correctly in the client area.

See #10