Fix gunshot light tracers not disappearing and refactor many effects classes
The ZEffectLightTracer::Draw function that was supposed to make it die was not getting called because a change in signature made it no longer an override of ZEffect::Draw.

Instead of changing it back to the u32 time signature, I've changed ZEffect::Draw to take u64 instead, and changed all of its derived classes in this way as well.

Additionally, I've:

- Added override specifiers to prevent this from happening again
- Changing some of the derived types' time members to be u64 to not have the new argument just be truncated
- Removed some empty, pointless destructors
- Removed ZEffectMesh.h and -.cpp since they didn't contain any uncommented code
- Reformatted and removed useless comments