Fix chat messages from previous game being visible when new game starts
The chat message time values were previously tied to the game time (ZGetGame::GetTime), which was reset back to zero when a new game starts. For instance, if you sent a message 100 seconds into a game, and then left that and started a new one, the chat message's time would still be set to 100 even though the game time is 0, and the message would be visible and non-faded for 100 seconds. It would only fade before that if a new message, sent in the new game, had faded, which would cause the code to fade all messages before it also, since it assumes the time values only increase.