Fix packets being delayed
There are two issues here being fixed here. The first issue, present in the original Gunz code, is that it would reset the send event after handling it, even though it already autoresets when WaitForMultipleEvents returns.

This could make for an up to around 100 ms long delay between two packets sent consecutively, with an order of events like:

* Main thread sends first packet, which pushes it onto the socket thread's send queue and sets the send event
* The socket thread wakes up and starts handling the packet in FlushSend (automatically resetting the send event when waking up)
* Main thread sends second packet, setting the send event again
* Socket thread returns from FlushSend, resetting the send event
* Socket thread goes back to waiting, but isn't woken up immediately
* WaitForMultipleEvents reaches the timeout value (100), waking up the thread

The second issue, which was added in RG only, is that the timeout was changed to infinite, which would make such packets stuck in the send queue until another packet is sent.